Expose UserType
